backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
cryptography==42.0.5
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pyotp==2.9.0
qrcode==7.4.2
scikit-learn==1.4.2
joblib==1.4.0
numpy==1.26.4
pandas==2.2.2
slowapi==0.1.9
pytest==8.1.1
pytest-asyncio==0.23.6
